+++ src/settings/cards/task/taskConfigCard.ts
@@ -5,12 +5,36 @@ import { ColorHelper } from "powerbi-visuals-utils-colorutils";
 
 import { formattingSettings } from "powerbi-visuals-utils-formattingmodel";
 import Card = formattingSettings.SimpleCard;
+import Group = formattingSettings.Group;
 import ColorPicker = formattingSettings.ColorPicker;
 import Slice = formattingSettings.Slice;
 import NumUpDown = formattingSettings.NumUpDown;
 import { ISetHighContrastMode } from "../interfaces/ISetHighContrastMode";
+import { CompositeCard } from "powerbi-visuals-utils-formattingmodel/lib/FormattingSettingsComponents";
 
-export class TaskConfigCardSettings extends Card implements ISetHighContrastMode {
+class BorderSettings extends Card {
+    public width = new NumUpDown({
+        name: "borderWidth",
+        displayNameKey: "Visual_Width",
+        value: 0,
+        options: {
+            minValue: {
+                type: ValidatorType.Min,
+                value: 0,
+            },
+            maxValue: {
+                type: ValidatorType.Max,
+                value: 5,
+            },
+        }
+    });
+
+    public slices: Slice[] = [this.width];
+    public name: string = "border";
+    public displayNameKey: string = "Visual_Border";
+}
+
+export class TaskConfigCardSettings extends CompositeCard implements ISetHighContrastMode {
     public static DefaultHeight: number = 40;
     public static MinHeight: number = 1;
 
@@ -34,14 +58,23 @@ export class TaskConfigCardSettings extends Card implements ISetHighContrastMode
         }
     });
 
+    public border = new BorderSettings();
+    public task = new Group({
+        name: "task",
+        displayNameKey: "Visual_Task",
+        slices: [this.fill, this.height],
+    });
+    public groups: Group[] = [ this.task, this.border];
+
+
     public name: string = "taskConfig";
     public displayNameKey: string = "Visual_TaskSettings";
-    public slices: Slice[] = [this.fill, this.height];
 
     public setHighContrastMode(colorHelper: ColorHelper): void {
         const isHighContrast = colorHelper.isHighContrast;
 
-        this.slices.forEach((slice) => {
+        const slices = this.groups.flatMap(group => group.slices);
+        slices.forEach((slice) => {
             if (slice instanceof ColorPicker) {
                 slice.value.value = colorHelper.getHighContrastColor("foreground", slice.value.value);
                 slice.visible = !isHighContrast;